# api/openapi.yaml
openapi: 3.1.0
info:
  title: LMS Service API
  version: 1.0.0
  description: Phase-1 LMS REST API. Timestamps are UTC and money uses integer minor units.
servers:
  - url: /
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/refresh:
    post: {operationId: refresh, tags: [Authentication], security: [], requestBody: {$ref: '#/components/requestBodies/RefreshToken'}, responses: {'200': {$ref: '#/components/responses/Tokens'}, '401': {$ref: '#/components/responses/Problem'}}}
...
  /api/v1/auth/logout-all:
    post: {operationId: logoutAll, tags: [Authentication], responses: {'204': {description: All sessions revoked}, '401': {$ref: '#/components/responses/Problem'}}}
...
  /api/v1/auth/change-password:
    post: {operationId: changePassword, tags: [Authentication], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChangePasswordRequest'}}}}, responses: {'204': {description: Password changed and sessions revoked}, '422': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/forgot-password:
    post: {operationId: forgotPassword, tags: [Authentication], security: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ForgotPasswordRequest'}, example: {email: student@example.com}}}}, responses: {'202': {description: Generic accepted response}, '429': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/reset-password:
    post: {operationId: resetPassword, tags: [Authentication], security: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResetPasswordRequest'}}}}, responses: {'204': {description: Password reset}, '401': {$ref: '#/components/responses/Problem'}}}
  /api/v1/auth/verify-email:
    post: {operationId: verifyEmail, tags: [Authentication], security: [], requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TokenRequest'}}}}, responses: {'204': {description: Email verified}, '401': {$ref: '#/components/responses/Problem'}}}
...
  /api/v1/notifications:
    get: {operationId: listNotifications, tags: [Notifications], parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}], responses: {'200': {$ref: '#/components/responses/List'}}}
...
  /api/v1/notifications/{id}/read:
    parameters: [{$ref: '#/components/parameters/ID'}]
    patch: {operationId: markNotificationRead, tags: [Notifications], responses: {'200': {$ref: '#/components/responses/OK'}}}
...
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer, bearerFormat: JWT, description: Short-lived LMS access JWT. Refresh tokens are accepted only by refresh/logout.}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
    Cursor: {name: cursor, in: query, schema: {type: string}, description: Opaque timestamp and UUID tie-breaker.}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 25}}
...
  requestBodies:
...
    RefreshToken: {required: true, content: {application/json: {schema: {type: object, required: [refresh_token], additionalProperties: false, properties: {refresh_token: {type: string, minLength: 32, maxLength: 256, writeOnly: true}}}}}}
  responses:
    OK: {description: Successful response, content: {application/json: {schema: {type: object, additionalProperties: true}}}}
    List: {description: Cursor-paginated list, content: {application/json: {schema: {$ref: '#/components/schemas/ListResponse'}}}}
...
    Problem: {description: RFC 7807 problem details, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Tokens: {description: Rotating access and refresh pair, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
  schemas:
...
    Timestamp: {type: string, format: date-time}
    Problem:
      type: object
      required: [type, title, status]
      example: {type: about:blank, title: Forbidden, status: 403, detail: active enrollment is required, request_id: 018f4d4e-8b57-7bf0-a250-3bb54f2ab31a}
      properties: {type: {type: string, format: uri-reference}, title: {type: string}, status: {type: integer}, detail: {type: string}, instance: {type: string}, request_id: {type: string}, errors: {type: array, items: {$ref: '#/components/schemas/FieldError'}}}
    FieldError: {type: object, required: [field, message], properties: {field: {type: string}, message: {type: string}}}
    ListResponse: {type: object, required: [items], properties: {items: {type: array, items: {}}, next_cursor: {type: string}}}
...
    ChangePasswordRequest: {type: object, required: [current_password, new_password], additionalProperties: false, properties: {current_password: {type: string, format: password, writeOnly: true}, new_password: {type: string, format: password, minLength: 12, maxLength: 1024, writeOnly: true}}}
    ForgotPasswordRequest: {type: object, required: [email], additionalProperties: false, properties: {email: {type: string, format: email}}}
    ResetPasswordRequest: {type: object, required: [token, new_password], additionalProperties: false, properties: {token: {type: string, minLength: 32, maxLength: 256, writeOnly: true}, new_password: {type: string, format: password, minLength: 12, maxLength: 1024, writeOnly: true}}}
    TokenRequest: {type: object, required: [token], additionalProperties: false, properties: {token: {type: string, minLength: 32, maxLength: 256, writeOnly: true}}}
    TokenPair: {type: object, required: [access_token, refresh_token, token_type, expires_at], properties: {access_token: {type: string, writeOnly: true}, refresh_token: {type: string, writeOnly: true}, token_type: {type: string, const: Bearer}, expires_at: {$ref: '#/components/schemas/Timestamp'}}}
...